prototype/requirements.txt
numpy
scipy
fastapi
uvicorn[standard]
python-multipart
astronomy-engine
httpx
async_lru
slowapi
sep-pjw
astropy